[OrganisePictures] Added script
Added script to organise pictures in Dropbox folder.

Pictures and video arrive in the `Camera Uploads` directory located at
root (which contains a weird unicode "non-breaking space" in filename
instead of a class ASCII space character) in a flat manner.

This script allows to move the files in a master picture directory,
sorted with a directory structure sorted in years and months.
This structure is easier to browse into and avoids to load thousands
of file when looking for a single picture.